Added OS call prompt
Fixed registration for the signaling task (regression - was returning null if the task was already registered)
Added new app service call to get the foreground visibility state.
Added conditional call to show the OS call prompt when the foreground is not visible.
Changed the Post method for the Voip Channel to allow for sync. running of delegates, since the VoipCall relay handing was done in a separate thread resulting in the background call being suspended before the call prompt is shown.

Test:
- Deploy the app to desktop and phone.
- Testing on desktop:
- Close the desktop client UI
- Call from phone
- On the desktop you will see the OS call prompt. Answering or rejecting the call should work
- Testing on phone:
- Repeat the desktop testing steps, replacing desktop with phone

Note:
- There seems to be some instability on audio path regardless if the call is answered with the OS prompt or the UI options.