Video window positioning
[Description]
The local feedback video window and the peer video window
could be better positioned, especially on the phone. I think
the local feedback window should simply overlapped the peer
video window. The peer video Window should take most of the
screen on the phone.

[Changes]
Removed the part where the video size gets fixed.

[Test Cases]
1. Make a video call via CB
2. Look at the peer video and make sure it takes as much
of the screen size as the video size permits.

Notes:
1.Tested on tablets and some phones (Lumia 635, 550, 1520)
2.The right part of the peer video goes beyond the screen.
This is a known issue and there is a separate bug to fix.